+++ app/components/post/comment/comment/CommentEditInput.tsx
@@ -2,11 +2,11 @@ import { useAtomValue } from "jotai";
 import styles from "./CommentInput.module.css";
 import { userIdAtom, userImageAtom } from "@/app/modules/atoms";
 import { useForm } from "react-hook-form";
-import { checkUserLoginStatus } from "@/app/modules/api";
 import { useState } from "react";
 import { LoginAlert } from "../@common/LoginAlert";
 import { Comment } from "@/app/modules/types";
 import { editComment } from "@/app/modules/api/comment";
+import { checkUserLoginStatus } from "@/app/modules/api/auth";
 
 interface CommentForm {
   userComment: string;
